fix: Split keyword chip components to satisfy detekt TooManyFunctions
Move LostAndFoundKeywordChip and its addable/deletable variants together
with the AnimatedKeywordChipFlowGroup helper into a dedicated
LostAndFoundKeywordChip.kt file so KeywordChipGroup.kt stays under the
detekt TooManyFunctions threshold.
